requirements.txt
Flask==2.0.2
gunicorn==20.1.0
line-bot-sdk==2.0.1
mecab-python3
